requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.11
langchain>=0.0.267
langchain>=0.0.267
streamlit>=1.27.0
streamlit-lottie>=0.0.5
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0